table: use optional in find_method()

+++ compiler/tests/str_gen_test.v
@@ -7,7 +7,7 @@ fn test_array_str() {
 	println(f)
 	//s := f.str()
 	//println(s)
-	n := [i64(1), 2, 3]
+	n := [1, 2, 3]
 	assert n.str() == '[1, 2, 3]'
 	println(n)  // make sure the array is printable
 	n2 := [4,5,6]
